requirements.txt
asgiref==3.8.1
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.3
dj-rest-auth==7.0.0
Django==5.0.7
django-allauth==65.2.0
django-cleanup==8.1.0
django-cors-headers==4.4.0
django-tinymce==4.1.0
django-unfold==0.34.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
google-auth==2.36.0
gunicorn==22.0.0
idna==3.10
packaging==24.1
pillow==10.4.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.9.0
python-dotenv==1.0.1
requests==2.32.3
rsa==4.9
sqlparse==0.5.1
urllib3==2.2.3
